Update organización funciones.
Se a actualizado como se distribuyen las funciones del videojuego.
