--- BBDD def/conexion.py
@@ -1,33 +1,0 @@
-import pymysql
-
-# Configuración de conexión
-db_config = {
-    'host': 'server-proyecto-2025.mysql.database.azure.com',
-    'user': 'Proyecto',
-    'password': 'Huevo19421942',
-    'database': 'siete_y_medio',
-    'port': 3306
-}
-
-def connect_to_database():
-    """Establece una conexión con la base de datos MySQL."""
-    try:
-        # Crear la conexión
-        connection = pymysql.connect(
-            host=db_config['host'],
-            user=db_config['user'],
-            password=db_config['password'],
-            database=db_config['database'],
-            port=db_config['port']
-        )
-        print("Conexión exitosa a la base de datos.")
-        return connection
-    except pymysql.MySQLError as e:
-        print(f"Error al conectar a la base de datos: {e}")
-        return None
-
-def close_connection(connection):
-    """Cierra la conexión con la base de datos."""
-    if connection:
-        connection.close()
-        print("Conexión cerrada.")
